Move gulp-plugins to separate directory
