+++ gulpfile.js
@@ -3,7 +3,7 @@ var shell = require('gulp-shell')
 var concat = require('gulp-concat')
 var cson = require('gulp-cson')
 var es = require('event-stream')
-var resolveTransformations = require('./scripts/resolve-transformations.js')
+var resolveTransformations = require('./scripts/plugins/resolve-transformations.js')
 var sass = require('gulp-sass')
 var watch = require('gulp-watch')
 var zip = require('gulp-zip')
